pyproject.toml
[tool.poetry]
name = "timewarrior-teamwork"
version = "0.1.0"
description = "TimeWarrior plugin for integration with Teamwork"
authors = ["Braden Mars <bradenmars@bradenmars.me>"]
packages = [
    { include = "twtw" }
]


[tool.poetry.scripts]
twtw = 'twtw.cli:cli'

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.23.0"
python-dotenv = "^0.13.0"
python-dateutil = "^2.8.1"
rich = "^2.2.6"
dpath = "^2.0.1"
click = "^7.1.2"

[tool.poetry.dev-dependencies]


[tool.dephell.main]
from = {format = "poetry", path="pyproject.toml"}
to = {format = "setuppy", path="setup.py"}

[build-system]
requires = ["poetry>=0.12", "setuptools!=50.0", "wheel"]
build-backend = "poetry.masonry.api"
